For armhf, allow using only manylinux python wheels
For amd64 and i386 (and, if available, arm64), Python provides manylinux
wheels, which are wheels with binary packages compiled against some
"baseline" glibc version. This is guaranteed to run across a wide number
of distro versions, and if the distro version being used isn't new
enough for the manylinux build environment that was used, pip will fall
back to compiling locally.

For armhf, however, this doesn't appear to be the case; there are some
wheels compiled for linux_armv7l, but there's no guarantee that such a
wheel will successfully run on Debian Bullseye. For grpcio-tools, this
is the issue. There is a linux_armv7l wheel available, but there is a
shared library _protoc in the package that requires a newer version of
glibc and libstdc++ than what is in Debian Bullseye. Our official builds
are fine because the underlying build machines are arm64 build machines
with armv8l architecture, and because there's no linux_armv8l wheel,
python will locally compile it.

Therefore, for armhf, allow installing only manylinux wheels, or, if
one doesn't exist, compile from source locally.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>