bgp: Add support for EdgeVpnAggregator in the BGP templates
In the large edge DC design, devices with device type EdgeVpnAggregator
are basically T2 devices. T1 devices that peer with EVAs should consider
those peer devices as basically T2 devices (i.e. spine routers).

Because of this, if this device is a LeafRouter, and the neighbor device
is a EdgeVpnAggregator, then use v4.leaf.spine/v6.leaf.spine template.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>