Merged PR 9207272: [dhcp_server] Set to build dhcp_server image in vs image in pr check
Why I did it
Currently in this repo would not build dhcp_server container image by default, which would cause that building issue for dhcp_server introduced by other modules cannot be noticed in time.
This PR is to set build dhcp_server container in vs image.

How I did it
Set to build dhcp_server in vs image

How to verify it
In checks of this PR, dhcp_server is built in vs image but not in other platform images.

Related work items: #26131144