Merged PR 9273198: Modify teamd_increase_retry_count.py for internal images
Modify teamd_increase_retry_count.py for internal images

In our internal images, we have special handling for BGP sessions on T0.
Specifically, in our internal images, the default BGP status is down.
However, on T0, BGP is brought up automatically. Because of that, when
this script is checking CONFIG_DB and APPL_DB to determine what BGP
sessions are up, the handling for this needs to be a bit different.
Specifically, assume BGP sessions are up unless `admin_status=down` is
explicitly seen.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

Related work items: #25654100