[docker-snmp] Provides a config file for the polling interval in sonic_ax_impl
Description:
By default, SONiC does not put any restriction on CPU usage for the containers in the system.
It caused the snmp container consumed 100% CPU usage.

A feasible approach is to define a configure file to adjust the CPU usage of snmp container
when it is required under specific scenarios.
The interval of the polling tasks in snmp process also affects the fundamental CPU usage consumed by the snmp process.
So the configure file also needs to provide the option to change the update intervals in snmp process.

Goal:
- Modify codes to support the features listed below.
 1. Restrict the CPU usage of snmp container through a configure file
 2. Adjust the polling intervals in snmp process through a configure file

What did I do:
 - Adding snmp_daemon_control.json to set "update_frequency"