Upgrade most of the Dockerfiles to do staged builds
On newer versions of Docker, only the buildkit builder is supported, and
cannot be disabled by setting DOCKER_BUILDKIT to 0. The side effect of
this is that the behavior of `--squash` is different (see
moby/moby#38903). This will result in the container sizes being
significantly higher.

To work around this, make all of our builds two-stage builds, with the
`--squash` flag entirely removed. The way this works is in the first
stage, whatever new files/packages need to be added are added (along
with files/packages that need to be removed). Then, in the second stage,
all of the files from the final state of the first stage are copied to
the second stage.

As part of this, also consolidate the container cleanup code into
`post_run_cleanup`, and remove it from the individual containers (for
consistency). Also experiment a bit with not needing to explcitly
install library dependencies, and let apt install it as necessary. This
will help during upgrades in the case of ABI changes for packages.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>