Fix the docker root cleanup and creation commands
The docker root cleanup is removing the contents of the docker root
directory we create from within a container. However, this container
isn't using the container registry variable, which menas it may fail
depending on the network environment. Fix this by prefixing the
container registry variable

The docker root directory creation is missing the `shell` at the
beginning, which means the directory doesn't actually get created. While
the docker command later will still create the directory automatically,
fix this and make sure it gets created here.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>