Set a limit of 20000 messages to be queued
In case rsyslog can't forward messages to the host's rsyslog server,
messages will be queued so that they can be sent out later. For this
queue, set a limit of 20000 messages so that rsyslog doesn't take too
much memory. Assuming each message is 512 bytes, the approximate maximum
additional memory usage is 10MB.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>