Update libnl3 to 3.7.0
This updates libnl3 to the version that is in Bookworm. There is one
patch still pending that is not present in 3.7.0. The functionality
provided by the patch has been upstreamed and is present in newer
versions of libnl3, but the implementation is slightly different.

When the update to newer versions of libnl3 happens and this
functionality is present, some code changes may be needed.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>