[dhcp] Fix that /etc/hosts is not appended with the same record
== Why I did it

When a dhcp server hands out a fully qualified hostname in option 12,
the previous sethostname script would keep appending to /etc/hosts.

Ticket: #20579
Alternative-PR: #20404

== How I did it

Replace DHCP sethostname and sethostname6 so that:

- hostname takes precedence if it is a FQDN
- if domain name is set and hostname is not FQDN, combine them
- if only hostname is set, merge it with a previously set domain name

Then set both the FQDN and the short hostname in /etc/hosts:

  127.0.0.1 shortname.domain.com shortname

And make sure that the old one is removed when the hostname is updated
through DHCP.

== How to verify it

$ reason=BOUND new_host_name=leaf1 sh -x sethostname
$ reason=BOUND new_host_name=leaf2 new_domain_name=domain.com sh -x sethostname
$ reason=BOUND new_host_name=leaf3.domain.com sh -x sethostname

Check that /etc/hosts and /etc/hostname are as expected.

== Description for the changelog

Rework how dhcp client sets hostname from dhcp host and domain options
and keep it from adding duplicates to /etc/hosts.