[installer] Use the Grub version comes from SONiC OS
- Why I did it
The Grub in ONiE OS is always out of date.
Reinstall ONiE and SONiC for upgrading Grub is quite complex and takes a lot of time.

- How I did it
Include Grub libraries to the installer.
Specify the directory of Grub libraries for grub-install.

- How to verify it
Install image from ONiE.
Check the Grub version on the banner of SONiC Grub menu.