Add `--protected-mode no` flag to redis-chassis
Redis-chassis doesn't bind to the localhost address intentionally.
Because of this, Redis blocks TCP connections that are connecting to
non-loopback interface addresses when there's no password set (which is
the case for us) unless `--protected-mode no` is set. This has already
been done for docker-database, but now needs to be done for
docker-sonic-vs following the upgrade to Bookworm.

Therefore, in the supervisord.conf file, add `--protected-mode no` to
the arguments passed in for redis-chassis.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>