[libyang bug] workaround leaf-list via uses bug in BGP route-map
There is a bug in libyang when attempting to use table `BGP_PEER_GROUP_AF`
and likely `BGP_NEIGHBOR_AF` in respect to `route_map_in` and `route_map_out`.

The issue is with the use of `leaf-list` specifically when it is pulled in
via a `uses` clause `bgpcmn:sonic-bgp-cmn-af`.

The error message resembles that of if a child is specified that isn't
recognized at all:
```
All Keys are not parsed in BGP_PEER_GROUP_AF
dict_keys(['default|PEERS|ipv4_unicast'])
exceptionList:["'route_map_in'"]
```

Moving the leaf-list to the parent rather than being imported
through the `uses` clause works around this issue.

Signed-off-by: Brad House (@bradh352)