flashrom build fix: incorrect tag name
flashrom recently started failing to build with the below error:
```
Cloning into 'flashrom-0.9.7'...
/sonic/src/flashrom/flashrom-0.9.7 /sonic/src/flashrom
fatal: 'tags/0.9.7' is not a commit and a branch 'flashrom-src' cannot be created from it
```
Nothing in sonic-buildimage has changed in relation to this so presumably
flashrom upstream renamed their tags.

This commit just fixes the formatting of the tag name to use the new format.

Signed-off-by: Brad House (@bradh352)