Add a service to sync the system time with the RTC
Currently, during runtime, because of the options that we start ntpd
with, ntpd will never tell the kernel that the time is synchronized.
This means that during runtime, the hardware clock will never get
updated with the current system clock, which means that if there's an
unexpected power loss/reboot, then the device will come up with an
incorrect time. (In our regular reboot scripts, we explicitly write the
system time to the hardware clock, so that is covered.)

To handle that, add a timer and service to update the hardware clock
with the current system time every hour. This makes sure that if the
device does get rebooted for whatever reason, then the time is close to
correct. The first run will happen 30 minutes after bootup.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>