libyang3-py3: libyang3 uses a python cffi bindings
libyang 1.0.73 provided python bindings via swig.  Those have
been removed and libyang3 now requires the use of the cffi bindings
which are in a separate repo.

The binding names themselves are actually different ('yang' vs 'libyang')
so both should be installable simultaneously.

This is built as a debian package rather than a wheel since it contains
binary elements.