update dependencies to include libyang3
libyang1 and libyang3 can coexist (except for the dev packages).

Even if a package doesn't directly depend on libyang, full
dependency tracking isn't smart enough to pull in deps of an
actual specified dependency, so we need to tag all users of
libyang as also a user of libyang3 (including python bindings
if they were previously specified).

In a future commit we will remove all libyang1 references.