Install dmidecode package for ARM platforms (#21577)
- Why I did it
dmidecode is a package required for arm platforms including Nvidia DPUs. Platform dump in the tech support bundle was missing the dmidecode output. This occurred because dmidecode was not installed on the DPUs.

- How I did it
Added dmidecode to the list of required packages for installation and removed it from the installation path for the amd64 architecture only.

- How to verify it
Verified tech support bundle after the fix, and it now includes the dmidecode data as expected.

Signed-off-by: ram25794 <ssingamala@nvidia.com>